refactor: enforce single ChatTab test naming
