+++ vite.config.js
@@ -13,6 +13,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: './tests/setup.js'
+    setupFiles: './tests/setup.js',
+    include: ['tests/**/*.test.jsx']
   }
 })
